rename renderer to printer
