+++ src/printer/plugins/NodeRefGroupPrintPlugin.ts
