+++ src/usecase/render/to-dot.ts
@@ -1,5 +1,5 @@
 import { convertToAST, GraphModel } from '../../model/index.js';
-import { stringify } from '../../renderer/index.js';
+import { stringify } from '../../printer/index.js';
 
 export function toDot(object: GraphModel): string {
   const ast = convertToAST(object);
